Move all connected components to containers directory
